Move transform and plugins back to Frost
