+++ src/Frost.hs
@@ -1,11 +1,20 @@
 {-# LANGUAGE TemplateHaskell #-}
 module Frost where
 
+import Frost.Plugin
+import Frost.TimestampPlugin
+import Frost.DefaultsMandatoryPlugin
+import Data.List (find)
+import Control.Monad
 import Polysemy
 import Polysemy.Input
 import Polysemy.Output
 import Polysemy.Error
-import Text.Pandoc
+import PolysemyContrib
+import Text.Pandoc hiding (trace)
+import Text.Pandoc.Extensions
+import Data.Map.Strict
+import Data.Traversable
 
 data DynamicError = PluginNotAvailable String | DynamicError String
   deriving (Eq, Show)
@@ -17,3 +26,30 @@ generateDocs :: ( Member (Input Pandoc) r
 generateDocs  transform = input >>= transform >>= output
 
 
+transform :: Member (Error DynamicError) r => [Plugin r] -> Pandoc -> Sem r Pandoc
+transform plugins (Pandoc meta blocks) = do
+  let plugin = (head plugins) -- TODO LOL, use all plugins, not one
+  newMeta <- addToMeta plugin meta
+  newBlocks <- traverse (extractFrostBlocks plugins) blocks
+  return $ Pandoc newMeta (join newBlocks)
+
+  where
+    extractFrostBlocks :: Member (Error DynamicError) r => [Plugin r] -> Block -> Sem r [Block]
+    extractFrostBlocks plugins = (\case
+        Para [Code ("",[],[]) name] -> do
+          let maybePlugin = find (\p -> "frost:" ++ pluginName p == name) plugins
+          case maybePlugin of
+            Just plugin ->  substitute plugin ""
+            Nothing -> throw $ PluginNotAvailable name
+        CodeBlock ("",[name],[]) content -> do
+          let maybePlugin = find (\p -> "frost:" ++ pluginName p == name) plugins
+          case maybePlugin of
+            Just plugin ->  substitute plugin content
+            Nothing -> throw $ PluginNotAvailable name
+        otherwise -> return [otherwise])
+
+plugins :: Member SystemEffect r => [Plugin r]
+plugins = [ timestampPlugin
+          , timestampMetaPlugin
+          , defaultsMandatoryPlugin
+          ]
